refactor: pre-split exports for sub-package migration
Extract shared types to protocol.go (ServerInfo, ClientInfo,
ClientCapabilities, StreamableHTTPAccept). Move AuthValidator/AuthError
to auth.go. Export ContextWithSession, MarshalNotification,
IsJSONRPCResponse, SendServerRequest, RouteServerResponse for use
by upcoming server/ and client/ sub-packages.

No behavior changes — all tests pass.
