+++ dispatch.go
@@ -103,34 +103,6 @@ type promptEntry struct {
 	handler PromptHandler
 }
 
-// ServerInfo identifies this MCP server in the initialize response.
-type ServerInfo struct {
-	Name         string `json:"name"`
-	Version      string `json:"version"`
-	Title        string `json:"title,omitempty"`
-	Description  string `json:"description,omitempty"`
-	Instructions string `json:"instructions,omitempty"`
-	WebsiteURL   string `json:"websiteUrl,omitempty"`
-}
-
-// ClientInfo identifies the MCP client from the initialize request.
-type ClientInfo struct {
-	Name    string `json:"name"`
-	Version string `json:"version"`
-}
-
-// ClientCapabilities describes features the client supports.
-type ClientCapabilities struct {
-	Sampling    *struct{} `json:"sampling,omitempty"`
-	Roots       *RootsCap `json:"roots,omitempty"`
-	Elicitation *struct{} `json:"elicitation,omitempty"`
-}
-
-// RootsCap describes the client's roots capability.
-type RootsCap struct {
-	ListChanged bool `json:"listChanged,omitempty"`
-}
-
 // initializeParams is the params object sent by the client in an initialize request.
 type initializeParams struct {
 	ProtocolVersion string             `json:"protocolVersion"`
