+++ elicitation_test.go
@@ -28,7 +28,7 @@ func TestElicitNotSupported(t *testing.T) {
 		t.Fatal("request func should not be called")
 		return nil, nil
 	})
-	ctx := contextWithSession(context.Background(), nil, request, &logLevel, caps, nil)
+	ctx := ContextWithSession(context.Background(), nil, request, &logLevel, caps, nil)
 
 	_, err := Elicit(ctx, ElicitationRequest{Message: "test"})
 	if err != ErrElicitationNotSupported {
